requirements.txt
# Runtime dependencies for Avis.
# Pin exact versions for reproducibility across machines.
# Update version pins when upgrading — don't leave ranges in production.
#
# Install: pip install -r requirements.txt

# ── Data validation / schema ─────────────────────────────────────────────────
pydantic>=2.7.4,<3.0.0           # BirdObservation and ClassificationResult schemas

# ── Config loading ────────────────────────────────────────────────────────────
pyyaml==6.0.1              # Load configs/*.yaml files

# ── Audio processing (Phase 2) ────────────────────────────────────────────────
librosa==0.10.2            # WAV loading, mel spectrogram generation
sounddevice==0.4.7         # Microphone capture (USB mic on Pi)
soundfile==0.12.1          # WAV file I/O (required by librosa on some platforms)

# ── Image processing (Phase 2) ───────────────────────────────────────────────
Pillow==10.3.0             # Image resizing and format conversion
numpy==1.26.4              # Array operations across all pipelines

# ── Machine learning (Phase 3 — baselines) ───────────────────────────────────
scikit-learn==1.8.0        # KNN and SVM baselines, metrics, preprocessing
matplotlib==3.10.8         # Confusion matrices, per-class F1 plots, k-selection curves
pandas==3.0.2              # Split CSV loading and experiment log (notebooks/results/)
scikit-image==0.24.0       #HOG (Histogram of Oriented Gradients) feature extraction for visuam SVM baseline

# ── Machine learning (Phase 4 — pretrained models) ───────────────────────────
torch==2.3.0               # PyTorch — model training and CPU inference
torchvision==0.18.0        # EfficientNet / MobileNet pretrained weights + transforms
timm==1.0.3                # Additional pretrained vision models (EfficientNet variants)

# ── BirdNET inference (Phase 4) ──────────────────────────────────────────────
birdnetlib==0.9.0          # Python wrapper for BirdNET-Analyzer
tensorflow-cpu==2.21.0     # TFLite runtime required by birdnetlib for BirdNET inference
resampy==0.4.3             # Audio resampling required by birdnetlib/librosa for MP3 files

# ── Environment variables ─────────────────────────────────────────────────────
python-dotenv==1.0.1       # Load .env file into os.environ

# ── Logging ───────────────────────────────────────────────────────────────────
# Standard library logging is used — no extra dependency needed.

# ── Hailo inference (Pi only — install separately, see docs/SETUP.md) ─────────
# hailort                   # Installed via Hailo's own installer, not pip


# ── CPU YOLO for bird-presence gate (Phase 8, Branch 2) ──────────────────────
# Runs YOLOv8s on CPU to decide whether motion-triggered frames contain a bird
# before invoking the species classifier. See src/vision/detector.py and
# docs/investigations/hailo-2026-04-22.md for the rationale (stock Hailo YOLO
# HEF does not produce usable detections; custom-compiled HEF deferred to
# Branch 5).
ultralytics>=8.4.0,<9.0.0


# Phase 6+ — Agentic LLM layer
langchain-core>=0.3.0,<2.0.0
langchain-google-genai>=2.0.0,<3.0.0
langgraph>=1.0.0,<2.0.0
langchain>=0.3.0,<2.0.0


# ─────────────────────────────────────────────────────────────────
# Layer 2 — labeling-assistant review UI (tools/labeler/ui/)
# ─────────────────────────────────────────────────────────────────
# Add these lines to requirements.txt. Versions chosen to match what
# the Layer 2 test suite was developed against (123 tests passing).
# If you bump these, re-run: pytest tests/labeler/ui/
 
# FastAPI: web framework for /api routes, page routes, image serving
fastapi==0.118.0
 
# Uvicorn: ASGI server. The [standard] extra adds httptools + uvloop
# for ~30% throughput on Linux (no effect on Windows but harmless).
uvicorn[standard]==0.32.0
 
# python-multipart: required by FastAPI to parse form data. We don't
# use forms today but FastAPI will warn at startup if missing.
python-multipart==0.0.18
 
# httpx: only used by the test suite (TestClient under the hood).
# Listed here because tests are run from the same venv as the app.
httpx==0.28.1
 